+++ crates/matrix-sdk/src/account.rs
@@ -698,7 +698,8 @@ impl Account {
         Ok(self.client.send(request).await?)
     }
 
-    /// Get the content of an account data event of statically-known type.
+    /// Get the content of an account data event of statically-known type, from
+    /// storage.
     ///
     /// # Examples
     ///
@@ -726,7 +727,7 @@ impl Account {
         get_raw_content(self.client.state_store().get_account_data_event_static::<C>().await?)
     }
 
-    /// Get the content of an account data event of a given type.
+    /// Get the content of an account data event of a given type, from storage.
     pub async fn account_data_raw(
         &self,
         event_type: GlobalAccountDataEventType,
@@ -779,6 +780,14 @@ impl Account {
                 }
             }
         }
+    }
+
+    /// Fetch an account data event of statically-known type from the server.
+    pub async fn fetch_account_data_static<C>(&self) -> Result<Option<Raw<C>>>
+    where
+        C: GlobalAccountDataEventContent + StaticEventContent,
+    {
+        Ok(self.fetch_account_data(C::TYPE.into()).await?.map(Raw::cast))
     }
 
     /// Set the given account data event.
@@ -862,12 +871,12 @@ impl Account {
 
         // We are fetching the content from the server because we currently can't rely
         // on `/sync` giving us the correct data in a timely manner.
-        let raw_content = self.fetch_account_data(GlobalAccountDataEventType::Direct).await?;
+        let raw_content = self.fetch_account_data_static::<DirectEventContent>().await?;
 
         let mut content = if let Some(raw_content) = raw_content {
             // Log the error and pass it upwards if we fail to deserialize the m.direct
             // event.
-            raw_content.deserialize_as::<DirectEventContent>().map_err(|err| {
+            raw_content.deserialize().map_err(|err| {
                 error!("unable to deserialize m.direct event content; aborting request to mark {room_id} as dm: {err}");
                 err
             })?
@@ -1081,21 +1090,22 @@ impl Account {
     pub async fn fetch_media_preview_config_event_content(
         &self,
     ) -> Result<Option<MediaPreviewConfigEventContent>> {
-        // First we check if there is avalue in the stable event
+        // First we check if there is a value in the stable event
         let media_preview_config =
-            self.fetch_account_data(GlobalAccountDataEventType::MediaPreviewConfig).await?;
+            self.fetch_account_data_static::<MediaPreviewConfigEventContent>().await?;
 
         let media_preview_config = if let Some(media_preview_config) = media_preview_config {
             Some(media_preview_config)
         } else {
             // If there is no value in the stable event, we check the unstable
-            self.fetch_account_data(GlobalAccountDataEventType::UnstableMediaPreviewConfig).await?
+            self.fetch_account_data_static::<UnstableMediaPreviewConfigEventContent>()
+                .await?
+                .map(Raw::cast)
         };
 
         // We deserialize the content of the event, if is not found we return the
         // default
-        let media_preview_config = media_preview_config
-            .and_then(|value| value.deserialize_as::<MediaPreviewConfigEventContent>().ok());
+        let media_preview_config = media_preview_config.and_then(|value| value.deserialize().ok());
 
         Ok(media_preview_config)
     }